+++ src/main/java/insane96mcp/buildersconstruct/BuildersConstruct.java
@@ -5,17 +5,15 @@ import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
-import slimeknights.tconstruct.tools.TinkerModifiers;
 
 @Mod(BuildersConstruct.MOD_ID)
 public class BuildersConstruct
 {
     public static final String MOD_ID = "buildersconstruct";
-    public static final String RESOURCE_PREFIX = MOD_ID + ":";
 
     public BuildersConstruct(FMLJavaModLoadingContext context) {
         IEventBus bus = context.getModEventBus();
-        bus.register(new TinkerModifiers());
+        bus.register(new BCModifiers());
 
         MinecraftForge.EVENT_BUS.register(this);
     }